refactor: move _nml_helpers functions into neuroml adapter
Move all helper functions (run_lems_example, plot_lems_comparison,
compare_traces, LEMS_EXAMPLES, etc.) from the standalone
_nml_helpers.py into tvbo/adapters/neuroml.py.

Update all 27 Ex*.qmd files to import from tvbo.adapters.neuroml
instead of _nml_helpers, removing the sys.path hack that caused
ModuleNotFoundError in test_docs.py.

- Delete docs/.../examples/_nml_helpers.py (736 lines)
- Append helpers to neuroml.py with underscore-prefixed imports
- Use lazy LEMS_EXAMPLES proxy to avoid git clone at import time
- All 27 NeuroML examples pass (PASS: 27, FAIL: 0, ERROR: 0)
