+++ docs/Interoperability/NeuroML/examples/Ex0_IaF.qmd
@@ -67,9 +67,7 @@ print(xml[:2000])
 ## 3. Run Reference (NeuroML2 via jNeuroML)
 
 ```{python}
-import sys, os
-sys.path.insert(0, os.path.dirname(os.path.abspath(".")))
-from _nml_helpers import run_lems_example, LEMS_EXAMPLES
+from tvbo.adapters.neuroml import run_lems_example, LEMS_EXAMPLES
 
 ref_outputs = run_lems_example("LEMS_NML2_Ex0_IaF.xml")
 print(f"Reference output files: {list(ref_outputs.keys())}")
@@ -88,6 +86,6 @@ print(f"TVBO: {da.dims}, shape={da.shape}")
 ## 5. Compare & Plot
 
 ```{python}
-from _nml_helpers import plot_lems_comparison
+from tvbo.adapters.neuroml import plot_lems_comparison
 plot_lems_comparison("LEMS_NML2_Ex0_IaF.xml", ref_outputs, result.integration.data, title_prefix="Ex0")
 ```
